fix docstring + moved some funs to legacy
